requirements.txt
torch
pickle
pillow
pymongo